requirements.txt
python-telegram-bot==20.7
requests==2.31.0
Pillow==10.2.0
flask==3.0.0
